refactor: decompose server/mod.rs into config, middleware, ws modules
